[Refactor] user graphql 분리 및 schema에 불필요한 코드 삭제
graphql resolver를 단위별로 분리 / 중복된 쿼리 제거 / requestDB를 사용하여 코드의 가독성 높임
