+++ server/src/api/auth/auth.mutation.ts
@@ -1,17 +1,11 @@
-import uploadToObjStorage from '../../middleware/uploadToObjStorage';
-import {
-  MutationSignUpArgs,
-  User,
-  MutationResolvers,
-  QueryResolvers
-} from '../../types';
 import { requestDB } from '../../utils/requestDB';
 import { encodeJWT } from '../../utils/jwt';
+import uploadToObjStorage from '../../middleware/uploadToObjStorage';
+import { MutationSignUpArgs, User, MutationResolvers } from 'src/types';
 import SameEmailError from '../../errors/EmailAlreadyExistsError';
 import { FIND_USER_BY_EMAIL_QUERY } from '../../schema/user/query';
-import { getUserWithStatus } from '../../schema/user/user';
 import { parseResultRecords } from '../../utils/parseData';
-import { loginPublish } from './user.pubsub';
+import { getNode } from '../../utils/parseDB';
 
 const checkIsEmailExist = async (email): Promise<void> => {
   const sameUsers = await requestDB(FIND_USER_BY_EMAIL_QUERY, { email });
@@ -40,13 +34,13 @@ const createUser = async info => {
     `CREATE (u:User {nickname: $nickname, hometown: $hometown, residence: $residence, email: $email ${
       info.thumbnail ? ', thumbnail: $thumbnail' : ''
     }
-      }) RETURN u`,
+        }) RETURN u`,
     info
   );
-  return result[0].get(0).properties;
+  return getNode(result);
 };
 
-const Mutation: MutationResolvers = {
+const mutationResolvers: MutationResolvers = {
   signUp: async (_, args: MutationSignUpArgs, { res }): Promise<User> => {
     await checkIsEmailExist(args.email);
     const thumbnail = await getUrlWhenFileExists(args.file);
@@ -57,15 +51,4 @@ const Mutation: MutationResolvers = {
   }
 };
 
-const Query: QueryResolvers = {
-  loginUser: async (_, __, { req }): Promise<boolean> => {
-    const user = await getUserWithStatus(req.email, 'online');
-    loginPublish(user);
-    return true;
-  }
-};
-
-export default {
-  Query,
-  Mutation
-};
+export default mutationResolvers;
