+++ server/src/api/auth/auth.resolvers.ts
@@ -1,0 +1,7 @@
+import Query from './auth.query';
+import Mutation from './auth.mutation';
+
+export default {
+  Query,
+  Mutation
+};
