+++ server/src/api/user/user.graphql
@@ -8,9 +8,8 @@ type Friend {
 type Query {
   me: User!
   getUser(email: String!): User
-  friends: [Friend!]!
-  userWithEmail(email: String!): User
-  friendsWithUserEmail(email: String!): [User]!
+  getFriends: [Friend!]!
+  getFriendsWithUserEmail(email: String!): [User]!
   findRelation(email: String!): String!
 }
