--- server/src/schema/user/type.ts
@@ -1,3 +1,0 @@
-export interface FindUserWithEmailQueryArgs {
-  email: string;
-}
