+++ server/src/utils/requestDB.ts
@@ -1,8 +1,9 @@
 import db from '../db';
 import createDBError from '../errors/createDBError';
 import { FIND_USER_BY_EMAIL_QUERY } from '../schema/user/query';
+import { getNode } from './parseDB';
 
-async function requestDB(query: string, param?) {
+export const requestDB = async (query: string, param?) => {
   let session;
   try {
     session = db.session();
@@ -15,11 +16,19 @@ async function requestDB(query: string, param?) {
   } finally {
     session.close();
   }
-}
+};
 
-async function getUserInfoByEmail(email: string) {
+export const getUserInfoByEmail = async (email: string) => {
   const user = await requestDB(FIND_USER_BY_EMAIL_QUERY, { email });
-  return user[0].get(0).properties;
-}
+  return getNode(user);
+};
 
-export { requestDB, getUserInfoByEmail };
+export const getUserWithStatus = async (email, status) => {
+  const result = await requestDB(FIND_USER_BY_EMAIL_QUERY, {
+    email
+  });
+  if (!result) return;
+  const user = getNode(result);
+  user.status = status;
+  return user;
+};
